Pass the span-name limit as int64 end to end
CodeQL flagged the int64-to-int conversion in getSpanNames; the clamp
above it makes overflow impossible, but the analyzer cannot see the range
through the branches, and the conversion bought nothing anyway -- the
decoder yields int64 and DuckDB binds it directly. No conversion, no
finding.

+++ desktopexporter/internal/store/spans/span_names_test.go
@@ -40,7 +40,7 @@ func namedSpans(names map[string]int) ptrace.Traces {
 	return tr
 }
 
-func spanNames(t *testing.T, s *store.Store, term string, limit int) []string {
+func spanNames(t *testing.T, s *store.Store, term string, limit int64) []string {
 	t.Helper()
 	raw, err := readStore(s, func(db *sql.DB) (json.RawMessage, error) {
 		return spans.GetSpanNames(t.Context(), db, term, limit)
